pyproject.toml
[tool.poetry]
name = "clickfile"
version = "0.1.0"
description = ""
authors = ["Weslly Oliveira <weslly99@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
boto3 = "^1.24.7"
django-storages = "^1.12.3"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest-django = "^4.5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
